Revert "✨ Promote Single Own Feature Gate AND Config spec in the CR to GA (OPRUN-4098) (#2268)"
This reverts commit 3c2fcb4c76acfea4f7ab5ba3b92adfdfecb23d57.
