+++ docs/tutorials/explore-available-content.md
@@ -91,6 +91,9 @@ Then you can query the catalog by using `curl` commands and the `jq` CLI tool to
         ...
         ```
 
+    !!! important
+        OLM 1.0 supports installing extensions that define webhooks. Targeting a single or specified set of namespaces requires enabling the `SingleOwnNamespaceInstallSupport` feature-gate.
+
 3. Return list of packages that support `AllNamespaces` install mode and do not use webhooks:
 
     ``` terminal
