+++ helm/tilt.yaml
@@ -14,6 +14,7 @@ options:
   operatorController:
     features:
       enabled:
+        - SingleOwnNamespaceInstallSupport
         - PreflightPermissions
         - HelmChartSupport
       disabled:
