+++ manifests/standard.yaml
@@ -613,45 +613,6 @@ spec:
             description: spec is an optional field that defines the desired state
               of the ClusterExtension.
             properties:
-              config:
-                description: |-
-                  config is an optional field used to specify bundle specific configuration
-                  used to configure the bundle. Configuration is bundle specific and a bundle may provide
-                  a configuration schema. When not specified, the default configuration of the resolved bundle will be used.
-
-                  config is validated against a configuration schema provided by the resolved bundle. If the bundle does not provide
-                  a configuration schema the final manifests will be derived on a best-effort basis. More information on how
-                  to configure the bundle should be found in its end-user documentation.
-                properties:
-                  configType:
-                    description: |-
-                      configType is a required reference to the type of configuration source.
-
-                      Allowed values are "Inline"
-
-                      When this field is set to "Inline", the cluster extension configuration is defined inline within the
-                      ClusterExtension resource.
-                    enum:
-                    - Inline
-                    type: string
-                  inline:
-                    description: |-
-                      inline contains JSON or YAML values specified directly in the
-                      ClusterExtension.
-
-                      inline must be set if configType is 'Inline'.
-                      inline accepts arbitrary JSON/YAML objects.
-                      inline is validation at runtime against the schema provided by the bundle if a schema is provided.
-                    type: object
-                    x-kubernetes-preserve-unknown-fields: true
-                required:
-                - configType
-                type: object
-                x-kubernetes-validations:
-                - message: inline is required when configType is Inline, and forbidden
-                    otherwise
-                  rule: 'has(self.configType) && self.configType == ''Inline'' ?has(self.inline)
-                    : !has(self.inline)'
               install:
                 description: |-
                   install is an optional field used to configure the installation options
